# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API for the owner-managed Project Hub catalog.
 * OpenAPI spec version: 0.1.0
 */
export type AuthUserRole = typeof AuthUserRole[keyof typeof AuthUserRole];


export const AuthUserRole = {
  user: 'user',
  admin: 'admin',
} as const;

export interface AuthUser {
  id: string;
  email: string;
  displayName: string;
  role: AuthUserRole;
}

export interface AuthSession {
  authenticated: boolean;
  user: AuthUser | null;
}

export interface AuthResult {
  user: AuthUser;
}

export interface LoginInput {
  /** @maxLength 254 */
  email: string;
  /**
     * @minLength 12
     * @maxLength 128
     */
  password: string;
}

export interface RegistrationInput {
  /** @maxLength 254 */
  email: string;
  /**
     * @minLength 1
     * @maxLength 100
     */
  displayName: string;
  /**
     * @minLength 12
     * @maxLength 128
     */
  password: string;
  /** @maxLength 512 */
  ownerSetupToken?: string;
}

export interface HealthStatus {
  status: string;
}

export type ProjectAccess = typeof ProjectAccess[keyof typeof ProjectAccess];


export const ProjectAccess = {
  free: 'free',
  premium: 'premium',
} as const;

export type ProjectStatus = typeof ProjectStatus[keyof typeof ProjectStatus];


export const ProjectStatus = {
  draft: 'draft',
  published: 'published',
  archived: 'archived',
} as const;

export interface Project {
  id: string;
  slug: string;
  title: string;
  shortDescription: string;
  description: string;
  category: string;
  tags: string[];
  access: ProjectAccess;
  status: ProjectStatus;
  featured: boolean;
  isNew: boolean;
  /** @nullable */
  thumbnailUrl?: string | null;
  /** @nullable */
  previewVideoUrl?: string | null;
  version: string;
  /** @nullable */
  fileSize?: string | null;
  views: number;
  downloads: number;
  favorites: number;
  releaseDate: string;
  updatedAt: string;
  screenshots: string[];
  supportedDevices: string[];
  requirements: string[];
  whatsNew: string;
  changelog: string[];
  fileAvailable: boolean;
  videoLocked: boolean;
}

export type ProjectDetail = Project & {
  isFavorite: boolean;
};

export type ProjectInputAccess = typeof ProjectInputAccess[keyof typeof ProjectInputAccess];


export const ProjectInputAccess = {
  free: 'free',
  premium: 'premium',
} as const;

export interface ProjectInput {
  /**
     * @minLength 2
     * @maxLength 100
     */
  slug: string;
  /**
     * @minLength 1
     * @maxLength 120
     */
  title: string;
  /**
     * @minLength 1
     * @maxLength 220
     */
  shortDescription: string;
  /**
     * @minLength 1
     * @maxLength 10000
     */
  description: string;
  /**
     * @minLength 1
     * @maxLength 60
     */
  category: string;
  /** @items.maxLength 40 */
  tags?: string[];
  access: ProjectInputAccess;
  featured?: boolean;
  isNew?: boolean;
  /** @nullable */
  thumbnailUrl?: string | null;
  /** @nullable */
  previewVideoUrl?: string | null;
  /** @maxLength 40 */
  version?: string;
  /** @nullable */
  fileSize?: string | null;
  releaseDate?: string;
  screenshots?: string[];
  supportedDevices?: string[];
  requirements?: string[];
  whatsNew?: string;
  changelog?: string[];
}

export type ProjectUpdateAccess = typeof ProjectUpdateAccess[keyof typeof ProjectUpdateAccess];


export const ProjectUpdateAccess = {
  free: 'free',
  premium: 'premium',
} as const;

export interface ProjectUpdate {
  /**
     * @minLength 1
     * @maxLength 120
     */
  title?: string;
  /**
     * @minLength 1
     * @maxLength 220
     */
  shortDescription?: string;
  /**
     * @minLength 1
     * @maxLength 10000
     */
  description?: string;
  /**
     * @minLength 1
     * @maxLength 60
     */
  category?: string;
  /** @items.maxLength 40 */
  tags?: string[];
  access?: ProjectUpdateAccess;
  featured?: boolean;
  isNew?: boolean;
  /** @nullable */
  thumbnailUrl?: string | null;
  /** @nullable */
  previewVideoUrl?: string | null;
  /** @maxLength 40 */
  version?: string;
  /** @nullable */
  fileSize?: string | null;
  releaseDate?: string;
  screenshots?: string[];
  supportedDevices?: string[];
  requirements?: string[];
  whatsNew?: string;
  changelog?: string[];
}

export interface PublicationInput {
  published: boolean;
}

export interface Account {
  userId: string;
  displayName: string;
  email: string;
  isAdmin: boolean;
  favoriteCount: number;
  downloadCount: number;
}

export interface ActionResult {
  success: boolean;
  /** @nullable */
  message?: string | null;
}

export interface DownloadResult {
  success: boolean;
  /** @nullable */
  downloadUrl: string | null;
  /** @nullable */
  message?: string | null;
}

export interface DownloadRecord {
  id: string;
  projectSlug: string;
  projectTitle: string;
  createdAt: string;
}

export interface Notification {
  id: string;
  title: string;
  message: string;
  createdAt: string;
  read: boolean;
}

export type SubscriptionStatus = typeof SubscriptionStatus[keyof typeof SubscriptionStatus];


export const SubscriptionStatus = {
  inactive: 'inactive',
  active: 'active',
  cancelled: 'cancelled',
  past_due: 'past_due',
} as const;

export interface Subscription {
  status: SubscriptionStatus;
  /** @nullable */
  plan: string | null;
  /** @nullable */
  expiresAt?: string | null;
  premiumEnabled: boolean;
}

export interface AdminOverview {
  projects: number;
  published: number;
  views: number;
  downloads: number;
  favorites: number;
  users: number;
  subscriptions: number;
  popular: Project[];
}

export type ListProjectsParams = {
q?: string;
category?: string;
access?: ListProjectsAccess;
sort?: ListProjectsSort;
featured?: boolean;
};

export type ListProjectsAccess = typeof ListProjectsAccess[keyof typeof ListProjectsAccess];


export const ListProjectsAccess = {
  all: 'all',
  free: 'free',
  premium: 'premium',
} as const;

export type ListProjectsSort = typeof ListProjectsSort[keyof typeof ListProjectsSort];


export const ListProjectsSort = {
  latest: 'latest',
  popular: 'popular',
} as const;